% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{body_fat}
\alias{body_fat}
\title{Simple body fat example data}
\format{
\subsection{\code{body_fat}}{

A data frame with 12 rows and 9 columns:
\describe{
\item{Subject}{Subject ID}
\item{group}{group}
\item{Weight_kg}{weight}
\item{Height_cm}{height}
\item{Age}{age}
\item{pct_body_fat}{percent body fat}
\item{survey_wt}{survey weight}
\item{stratum}{stratum}
\item{cluster}{cluster}

}
}
}
\usage{
body_fat
}
\description{
Simple body fat example data
}
\keyword{datasets}
