% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveynnet.R
\name{surveynnet}
\alias{surveynnet}
\title{Title}
\usage{
surveynnet(x, y, weight, size = 3, maxit = 20000, strat, clust)
}
\arguments{
\item{x}{blch}

\item{y}{asdf}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
none
}
