% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveynnet.R
\name{surveynnet}
\alias{surveynnet}
\title{Title}
\usage{
surveynnet(x, y, weight, size = 3, maxit = 20000, strat, clust)
}
\arguments{
\item{x}{matrix or data frame of predictors}

\item{y}{vector of targets / response values}

\item{weight}{weights for each sample}

\item{size}{size for number of units in the hidden layer of nnets}

\item{maxit}{maximum number of iterations, default 20000}
}
\value{
a dataframe
}
\description{
Title
}
\examples{
none
}
