% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveynnet.R
\name{surveynnet}
\alias{surveynnet}
\title{Neural Net for Complex Survey Data}
\usage{
surveynnet(x, y, weight, strat, clust, ...)
}
\arguments{
\item{x}{Matrix or data frame of predictors. Must not contain any missing values.}

\item{y}{Vector of targets / response values. Must not contain any missing values.}

\item{weight}{The weights for each sample.}

\item{strat}{The stratum for each sample.}

\item{clust}{The cluster for each sample.}

\item{...}{Additional arguments to be passed into \code{PracTools::deffCR} or \code{nnet::nnet}. See
documentation of those packages and functions for more details. Note that for the neural net (\code{nnet}),
the default here is set to 3 layers ("size" parameter) and maximum iterations ("maxit" parameter) is
set to 2000.}
}
\value{
A dataframe containing the results of the neural net, using:
\itemize{
\item no weights
\item the user-inputted weights
\item the new method that adjusts the weights by using a design effect incorporating cluster and strata
}
}
\description{
The surveynnet package extends the functionality of nnet (Venables and Ripley, 2002),
which already supports survey weights, by enabling it to handle clustered and stratified data.
It achieves this by incorporating design effects through the use of effective sample sizes in
the calculations, performed by the package described in Valliant et al. (2023), by following
the methods outlined by Chen and Rust (2017) and Valliant et al. (2018).
}
\examples{

# short example with body fat dataset
y <- body_fat$pct_body_fat
x <- body_fat[,c("Weight_kg", "Height_cm", "Age")]
weight <- body_fat$survey_wt
strat <- body_fat$stratum
clust <- body_fat$cluster
y[strat==1] <- y[strat==1] + 30*0.00015*rnorm(sum(strat==1))
y[strat==2] <- y[strat==2] + 30*0.15*rnorm(sum(strat==2))
myout <- surveynnet(x,y,weight = weight, strat = strat, clust=clust)
myout
}
\references{
\itemize{
\item Venables WN, Ripley BD (2002). Modern Applied Statistics with S, Fourth edition.
Springer, New York. ISBN 0-387-95457-0, https://www.stats.ox.ac.uk/pub/MASS4/.
\item Chen, S., and K. F. Rust. 2017."An Extension of Kish’s Formula for Design Effects to
Two- and Three-Stage Designs with Stratification.”, Journal of Survey Statistics and
Methodology,5 (2): 111–30.
\item Valliant, R., J. A. Dever, and F. Kreuter. 2018. Practical Tools for Designing
and Weighting Survey Samples .2nd ed. New York: Springer-Verlag.
\item Valliant, R., J. A. Dever, and F. Kreuter. 2023. PracTools: Tools for Designing
and Weighting Survey Samples, Version 1.4 . https://CRAN.R-project.org/package=PracTools
}
}
