% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveynnet.R
\name{surveynnet}
\alias{surveynnet}
\title{Title}
\usage{
surveynnet(x, y, weight, strat, clust, ...)
}
\arguments{
\item{x}{Matrix or data frame of predictors. Must not contain any missing values.}

\item{y}{Vector of targets / response values. Must not contain any missing values.}

\item{weight}{The weights for each sample.}

\item{strat}{The stratum for each sample.}

\item{clust}{The cluster for each sample.}

\item{...}{Additional arguments to be passed into \code{PracTools::deffCR} or \code{nnet::nnet}. See
documentation of those packages and functions for more details. Note that for the neural net (\code{nnet}),
the default here is set to 3 layers ("size" parameter) and maximum iterations ("maxit" parameter) is
set to 2000.}
}
\value{
A dataframe containing the results of the neural net, using:
\itemize{
\item no weights
\item the user-inputted weights
\item the new method that adjusts the weights by using a design effect incorporating cluster and strata
}
}
\description{
Title
}
\examples{

small example
}
\references{
Cruz-Cano, Cohen etc etc
}
